// web/generated/api.ts
export interface CreateProductTypeDto {
    /**
     * 
     * @type {string}
     * @memberof CreateProductTypeDto
     */
    'name': string;
    /**
     * 
     * @type {Array<CreateProductTypeDtoAttributes>}
     * @memberof CreateProductTypeDto
     */
    'attributes': Array<CreateProductTypeDtoAttributes>;
}
/**
 * 
 * @export
 * @interface CreateProductTypeDtoAttributes
 */
export interface CreateProductTypeDtoAttributes {
    /**
     * 
     * @type {string}
     * @memberof CreateProductTypeDtoAttributes
     */
    'name'?: string;
    /**
     * 
     * @type {Enum}
     * @memberof CreateProductTypeDtoAttributes
     */
    'type'?: Enum;
}
// ...
export interface UpdateProductTypeDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateProductTypeDto
     */
    'name'?: string;
    /**
     * 
     * @type {Array<CreateProductTypeDtoAttributes>}
     * @memberof UpdateProductTypeDto
     */
    'attributes'?: Array<CreateProductTypeDtoAttributes>;
}
// ...
export class ProductTypeApi extends BaseAPI {
    /**
     * 
     * @param {CreateProductTypeDto} createProductTypeDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerCreate(createProductTypeDto: CreateProductTypeDto, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerCreate(createProductTypeDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerFindAll(options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerFindAll(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerFindOne(id: string, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerFindOne(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerRemove(id: string, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerRemove(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {UpdateProductTypeDto} updateProductTypeDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductTypeApi
     */
    public productTypeControllerUpdate(id: string, updateProductTypeDto: UpdateProductTypeDto, options?: AxiosRequestConfig) {
        return ProductTypeApiFp(this.configuration).productTypeControllerUpdate(id, updateProductTypeDto, options).then((request) => request(this.axios, this.basePath));
    }
}